Update meal plan generation and schema enhancements
- Changed AI model from "gpt-4" to "gpt-5-mini".
- Introduced `PersistPlanActivity` for meal plan persistence.
- Fixed `MaxRecipes` parameter handling in recipe retrieval.
- Updated `EstimatedCost` type from `decimal` to `int` in `MealPlan`.
- Enhanced `MealPlanSchema.json` with new properties and requirements.
- Commented out recipe retrieval logic in `GenerateWeeklyPlanOrchestrator`.
- Improved prompt generation in `OpenAIService` to reflect user preferences.
- Added `maxCaloriesPerMeal` parameter in `SearchService`.
- Changed Azure Search index name to `reciepe-index-new`.
- Updated mock recipe data for consistency.
- Modified `local.settings.json` for new configurations.

+++ backend-api/dotnet/NutriMind.Api/NutriMind.Api/Models/MealPlan.cs
@@ -191,7 +191,7 @@ namespace NutriMind.Api.Models
         public string Category { get; set; } = string.Empty; // produce, dairy, meat, etc.
 
         [JsonProperty("estimatedCost")]
-        public decimal EstimatedCost { get; set; }
+        public int EstimatedCost { get; set; }
 
         [JsonProperty("isOptional")]
         public bool IsOptional { get; set; } = false;
